services/user_service/requirements.txt
FLASK
bcrypt
psycopg2-binary
PyJWT